Update link reference and revise color handling
Link reference on toner list page was adjusted to correctly point to 'toner_views.new_toner'. Color handling was also revised in this commit: the 'hex' attribute has been removed from ColorForm class and the new color form in the template, which simplifies the process of creating a new color. Additional changes made to the color list layout provide an intuitive user interface for creation and deletion of color entries.